perf: single-RPC admin stats and server-rendered receptionist dashboard
PERF-LAT-06 (admin dashboard): getDashboardStats issued 9 PostgREST
queries per render — five head-only counts plus three UNBOUNDED row
fetches (every completed payment, every review, every patient row with
metadata) just to compute sums/averages/counts in JS. New migration 00180
adds get_clinic_dashboard_stats(cid), a SECURITY INVOKER SQL function
(RLS still applies) returning all eight aggregates in one round trip.
The loader tries the RPC first and falls back to the legacy multi-query
path until the migration is applied (same pattern as avg_clinic_rating).
EXECUTE revoked from PUBLIC/anon, granted to authenticated.

PERF-LAT-07 (receptionist dashboard): the page was fully client-side —
it shipped a spinner, hydrated, ran getCurrentUser() (a browser-to-
Supabase auth round trip) and then three more fetches, a complete
post-hydration waterfall on every visit. Split into:
  - getReceptionistDashboardData(): server loader, one parallel query
    batch (appointments 30-day lookback + future, patients, doctor and
    service name lookups, completed payments in window)
  - ReceptionistDashboardView: client component keeping all interactivity
    (tabs, check-in state, optimistic updates, dialogs, realtime waiting
    room), now fed by props
  - page.tsx: thin Server Component using requireAuth (which since
    PERF-LAT-02 reuses the middleware's signed profile headers)
The existing (receptionist)/loading.tsx skeleton becomes the streaming
fallback, so users see a structured skeleton then the finished dashboard,
with zero client data round trips.

Verified: tsc --noEmit clean, full vitest suite passes (164 files,
1981 tests), eslint 0 errors, prettier clean.
